+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - [Optimize:optimize spring annotation.](https://github.com/Tencent/spring-cloud-tencent/pull/689)
 - [docs:update PR template.](https://github.com/Tencent/spring-cloud-tencent/pull/692)
 - [Code optimize & add junit tests.](https://github.com/Tencent/spring-cloud-tencent/pull/702)
+- [Test:support environment variable metadata test.](https://github.com/Tencent/spring-cloud-tencent/pull/711)